Move all type definitions out of src/types/index.ts
Move breadcrumbs and passport definitions into session.ts
